maint: add legacy_mode_v2 flag for default fixes
This adds usage of a new flag to `Network` named `legacy_mode_v2`. This
is intended as a compatibility flag that is distinct from the existing
`legacy_mode`. When true, this new flag uses the old
`drive["event_seed"]` behavior and the old synaptic delays.

I've thought about this and tried a couple of different ways to approach
this problem, and this makes the most sense to me. Doing it this way is
ugly, yes, but:
1. Backwards-compatibility is more important in scientific code than
usual. Adding various flags to support this is a small price to pay; the
cyclomatic complexities of the affected functions are already very high
quite frankly, but keeping the old logic around helps with
reproducibility.
2. Specifying more legacy flags is ugly, yes, but it also allows us to
heavily document what the *differences* between those versions are.
3. Doing it this way allows for a user to use different combinations of
different legacy modes. I initially tried changing `legacy_mode` to an
integer in order to turn it into a kind of "version", but then I
realized that this means the "versions" would be mutually exclusive,
which is bad. It is possible that someone will want to use all the
available legacy versions.
4. I think this is a forward-compatible solution as well. As long as we
are documenting everything well, this allows us to apply updates that
impact the *actual scientific output* of our default model (instead of
just file formats, etc.).

Note: for debug reasons, this flag is currently set to TRUE. But that's
just because I want to package this commit separately from the required
updates to the tests that this needs.

This is incomplete, since I still need to fix the tests that are broken
when the flag is off, and we need to discuss whether this is how we want
to move forward.