fix: update default networks
It turns out the remaining test failures were only due to
test-network-file incompabilities, and they needed to be regenerated.

Very interestingly: even though this changes the drive event seeds and
three affected synaptic delays in the default model, all tests
pass. This means that these changes do NOT affect the actual Dipole
output data from the default simulation; if it did, then this test
should fail:
https://github.com/jonescompneurolab/hnn-core/blob/90f4d636b13647869f11825daf40eb985b6965cb/hnn_core/tests/test_parallel_backends.py#L371

@ntolley @katduecker @dylansdaniels Do we expect these changes in the
default model to have literally no impact on the dipole output? That
seems surprising to me...

TODO: I have also not added *new* tests for the new flag, in part
because we still need to discuss if this is the solution we want to use.